requirements.txt
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
gcloud==0.17.0
googleapis-common-protos==1.5.3
gunicorn==19.8.1
httplib2==0.11.3
jws==0.1.3
oauth2client==3.0.0
oauthlib==2.1.0
protobuf==3.5.2.post1
psycopg2==2.7.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycryptodome==3.4.3
PyJWT==1.6.3
Pyrebase==3.0.27
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2018.3
requests==2.11.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
whitenoise==3.3.1
